Corrección de métodos.

+++ src/calculadora/Producto.java
@@ -54,7 +54,7 @@ public class Producto {
 	 * Si el valor introducido es negativo se devuelve código de error -1.
 	 * 
 	 */
-	public static double productoEntero(int num1, int num2) {
+	public static int productoEntero(int num1, int num2) {
 		if(num1 < 0 || num2 < 0)
 			return -1;
 		else
